+++ test/framework/chaos/podchaos/podchaos.go
@@ -1,4 +1,4 @@
-package chaos
+package podchaos
 
 import (
 	"context"
@@ -8,27 +8,25 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 	runtimeClient "sigs.k8s.io/controller-runtime/pkg/client"
 )
 
-type podChaos = *chmv1alpha1.PodChaos
-type PodSelector = *chmv1alpha1.PodSelector
-
-type PodChaos struct {
-	podChaos
-}
+type PodChaos chmv1alpha1.PodChaos
+type PodSelector = chmv1alpha1.PodSelector
 
 const (
-	podChaosCRDName string = "podchaos.chaos-mesh.org"
-	requiredVersion string = "v1alpha1"
+	CRDName         string = "podchaos.chaos-mesh.org"
+	RequiredVersion string = "v1alpha1"
 
 	PodKillAction       string = "pod-kill"
 	PodFailureAction    string = "pod-failure"
 	ContainerKillAction string = "container-kill"
 )
 
-func newPodChaos(namespace string, selector chmv1alpha1.PodSelector, opts ...func(podChaos)) PodChaos {
-	podChaos := &chmv1alpha1.PodChaos{
+// New returns a PodChaos object configured with a selector and provided options.
+func New(namespace string, selector *PodSelector, opts ...func(*PodChaos)) PodChaos {
+	podChaos := chmv1alpha1.PodChaos{
 		ObjectMeta: metav1.ObjectMeta{
 			Name:      "pod-failure",
 			Namespace: namespace,
@@ -36,25 +34,28 @@ func newPodChaos(namespace string, selector chmv1alpha1.PodSelector, opts ...fun
 		Spec: chmv1alpha1.PodChaosSpec{
 			Action: chmv1alpha1.PodFailureAction,
 			ContainerSelector: chmv1alpha1.ContainerSelector{
-				PodSelector: selector,
+				PodSelector: *selector,
 			},
 		},
 	}
 
+	podChaosObj := PodChaos(podChaos)
 	for _, lambda := range opts {
-		lambda(podChaos)
+		lambda(&podChaosObj)
 	}
 
-	return PodChaos{podChaos}
+	return podChaosObj
 }
 
-func withName(name string) func(podChaos) {
-	return func(c podChaos) {
+// WithName overrides the Name field for a PodChaos object.
+func WithName(name string) func(*PodChaos) {
+	return func(c *PodChaos) {
 		c.ObjectMeta.Name = name
 	}
 }
 
-func withPodFailureAction(action string) func(podChaos) {
+// WithAction overrides the PodChaos Action field.
+func WithAction(action string) func(*PodChaos) {
 	var a chmv1alpha1.PodChaosAction
 	switch action {
 	case PodKillAction:
@@ -67,11 +68,12 @@ func withPodFailureAction(action string) func(podChaos) {
 		panic("Invalid PodChaosAction : " + action)
 	}
 
-	return func(c podChaos) {
+	return func(c *PodChaos) {
 		c.Spec.Action = a
 	}
 }
 
+// CheckChaosActive checks if a PodChaos object indicates a successful injection of Chaos action.
 func (pc PodChaos) CheckChaosActive(ctx context.Context, c runtimeClient.Client) (bool, error) {
 	podChaos := &chmv1alpha1.PodChaos{}
 	err := c.Get(ctx, types.NamespacedName{Name: pc.Name, Namespace: pc.Namespace}, podChaos)
@@ -89,15 +91,15 @@ func (pc PodChaos) CheckChaosActive(ctx context.Context, c runtimeClient.Client)
 	return false, nil
 }
 
-func (pc PodChaos) Delete(ctx context.Context, c runtimeClient.Client) error {
-	if err := c.Delete(ctx, pc.podChaos); err != nil {
-		return fmt.Errorf("failed to delete PodChaos %s: %w", pc.Name, err)
-	}
-	return nil
+// GetObject returns the actual PodChaos object
+func (pc PodChaos) KubernetesObject() client.Object {
+	chaosObj := chmv1alpha1.PodChaos(pc)
+	return &chaosObj
 }
 
-func newPodLabelSelector(labels map[string]string,
-	opts ...func(PodSelector)) PodSelector {
+// NewPodLabelSelector returns a new PodSelector configured using labels and provided options.
+func NewPodLabelSelector(labels map[string]string,
+	opts ...func(*PodSelector)) *PodSelector {
 
 	podSelector := &chmv1alpha1.PodSelector{
 		Selector: chmv1alpha1.PodSelectorSpec{
@@ -115,7 +117,8 @@ func newPodLabelSelector(labels map[string]string,
 	return podSelector
 }
 
-func withSelectorMode(mode string) func(PodSelector) {
+// WithSelectorMode overrides the SelectorMode for a PodSelector.
+func WithSelectorMode(mode string) func(*PodSelector) {
 	var m chmv1alpha1.SelectorMode
 	switch mode {
 	case "one":
@@ -130,13 +133,14 @@ func withSelectorMode(mode string) func(PodSelector) {
 		m = chmv1alpha1.RandomMaxPercentMode
 	}
 
-	return func(s PodSelector) {
+	return func(s *PodSelector) {
 		s.Mode = m
 	}
 }
 
-func withSelectorNamespace(namespaces []string) func(PodSelector) {
-	return func(s PodSelector) {
+// WithSelectorNamespace overrides the namespaces for a PodSelector.
+func WithSelectorNamespace(namespaces []string) func(*PodSelector) {
+	return func(s *PodSelector) {
 		s.Selector.Namespaces = namespaces
 	}
 }